requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.5.2
joblib==1.4.2
plotly==5.24.1
python-dateutil==2.9.0.post0
holidays==0.57

